Add files via upload
int8_t changed to int16_t for bin_to_dec

+++ Task1.6/task1.6.c
@@ -51,8 +51,8 @@ void get_virtual_page(char vpn[6], char offset[9])                  // Asking fo
 
     do {
         printf("Enter 8-bit Offset value (eg: 01101111): ");
-        scanf("%8s", offset);                                           // Take in the first 8 input keyed in
-        clear_buffer();                                                 
+        scanf("%8s", offset);                                           // Take in the first 8 input keyed in       
+        clear_buffer();                                     
         if (strlen(offset) != 8) {                                      // Perform content length check
             printf("Invalid length, please enter exactly 8 bits.\n");   // Failed, print error message
             continue;                                                   // Immediately proceed to ask user for new input without performing content check
@@ -62,7 +62,7 @@ void get_virtual_page(char vpn[6], char offset[9])                  // Asking fo
 
 }
 
-int8_t bin_to_dec(const char *bitstring)                                // Function to convert binary to decimal
+int16_t bin_to_dec(const char *bitstring)                                // Function to convert binary to decimal
 {
     return (int)strtol(bitstring, NULL, 2);                             // Produce the result of computation of conversion
 }
